# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a demo version of Gin app.",
        "title": "Gin Demo App",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/arena": {
            "put": {
                "description": "Update a Arena by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arena"
                ],
                "summary": "Update a Arena",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update request payload",
                        "name": "updateReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateArenaReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arena updated successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "404": {
                        "description": "Arena not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new arena to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arena"
                ],
                "summary": "Add a new arena",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Arena request payload",
                        "name": "garageReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddArenaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arena added successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Arena by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arena"
                ],
                "summary": "Delete a Arena",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Arena request payload",
                        "name": "ArenaReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeletArenaReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arena deleted successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "404": {
                        "description": "Arena not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
...
        "/arena/get": {
            "get": {
                "description": "Retrieve the list of all Arenas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arena"
                ],
                "summary": "Get All Arena List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default is 0)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of records to fetch (default is 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arena list fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
        "/arena/types": {
            "get": {
                "description": "Retrieve the list of all arena types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arena"
                ],
                "summary": "Get All Arena type List",
                "responses": {
                    "200": {
                        "description": "Arena type list fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.AddArenaRequest": {
            "type": "object",
            "properties": {
                "arenaLevel": {
                    "type": "integer"
                },
                "arenaName": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "request.DeletArenaReq": {
            "type": "object",
            "properties": {
                "arenaId": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateArenaReq": {
            "type": "object",
            "properties": {
                "arenaId": {
                    "type": "string"
                },
                "arenaLevel": {
                    "type": "integer"
                },
                "arenaName": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "response.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}